server/requirements.txt
############################
# Core Web Backend (Flask)
############################
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
gunicorn==21.2.0
python-dotenv==1.0.1
requests==2.32.3

############################
# Database: PostgreSQL + PostGIS + ORM
############################
SQLAlchemy==2.0.30
GeoAlchemy2==0.15.2
psycopg2-binary==2.9.9
alembic==1.13.2

############################
# Caching / Background Jobs
############################
redis==5.0.4
celery==5.3.6

############################
# Cloud Storage: Google Cloud Storage
############################
google-cloud-storage==2.16.0

############################
# Authentication & Authorization
############################
Authlib==1.3.1
PyJWT==2.8.0
flask-jwt-extended==4.6.0
passlib[bcrypt]==1.7.4

############################
# Firebase (FCM + Auth)
############################
firebase-admin==6.5.0

############################
# AWS (Cognito, S3, etc.)
############################
boto3==1.34.134

############################
# Mapping & GIS (OpenStreetMap, Geo stuff)
############################
geopy==2.4.1
osmnx==1.9.0
Shapely==2.0.4
pyproj==3.6.1

############################
# Data Analytics & ML
############################
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
tensorflow==2.16.1
torch==2.3.0
pyspark==3.5.1

############################
# Observability (Logging)
############################
loguru==0.7.2

############################
# Dev Tools (optional)
############################
black==24.4.2
flake8==7.0.0
ipython==8.25.0
